# ndc-aggregator/server/docs/swagger.yaml
openapi: 3.0.3
info:
  title: IATA Digital Identity B2C PoC — NDC Aggregator API
  description: |
    REST API for the NDC Aggregator Proof-of-Concept.
    Integrates IATA Digital Identity (Verifiable Credentials via OpenID4VP)
    with multi-airline NDC capabilities (AirShopping, OfferPrice, OrderCreate).

    **Authentication** is handled via two mutually exclusive header patterns:
    - `X-Agency-Token` — session token from agency login (UC1), triggers server-side VP generation
    - `X-Verification-Id` — session ID from QR-based verification (UC2)
  version: 1.0.0
  contact:
    name: Dreamix
servers:
  - url: http://localhost:3000/api
    description: Local development
...
paths:
...
  /orders:
    post:
      tags: [NDC Orders]
      summary: Create an order (book)
      description: |
        Sends an OrderCreate request to the airline for the selected offer.
        Stores the order in the in-memory order store.
        In demo mode, returns a mock confirmed order.
      security:
        - AgencyToken: []
        - VerificationId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Missing required fields
        '401':
          description: Authentication failed

    get:
      tags: [NDC Orders]
      summary: List orders for current session
      security:
        - VerificationId: []
      responses:
        '200':
          description: Orders list
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrderSummary'
        '401':
          description: Authentication failed
  /orders/{orderId}:
    get:
      tags: [NDC Orders]
      summary: Get order details
      security:
        - AgencyToken: []
        - VerificationId: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Order not found

  # ── Agent ───────────────────────────────────────────────
  /me:
    get:
      tags: [Agent]
      summary: Get current agent profile
      description: Returns the agent profile extracted from the verification session.
      security:
        - VerificationId: []
      responses:
        '200':
          description: Agent profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentProfile'
        '404':
          description: Agent profile not found in session
components:
  securitySchemes:
    AgencyToken:
      type: apiKey
      in: header
      name: X-Agency-Token
      description: Session token from POST /auth/login (UC1)
    VerificationId:
      type: apiKey
      in: header
      name: X-Verification-Id
      description: Session ID from QR verification (UC2)
  schemas:
...
    Segment:
      type: object
      properties:
        flightNumber:
          type: string
          example: TK1979
        departureAirport:
          type: string
          example: IST
        departureTime:
          type: string
          format: date-time
        departureStation:
          type: string
        departureTerminal:
          type: string
        arrivalAirport:
          type: string
          example: LHR
        arrivalTime:
          type: string
          format: date-time
        arrivalStation:
          type: string
        arrivalTerminal:
          type: string
        duration:
          type: string
          example: PT4H10M
        carrierCode:
          type: string
        operatingCarrier:
          type: string
        aircraft:
          type: string
        aircraftName:
          type: string
        journeyDirection:
          type: string
          enum: [outbound, inbound]
    Price:
      type: object
      properties:
        total:
          type: number
          example: 245.50
        currency:
          type: string
          example: USD
        base:
          type: number
          nullable: true
        tax:
          type: number
          nullable: true
...
    OrderRequest:
      type: object
      required: [searchId, offerId, passengers]
      properties:
        searchId:
          type: string
          format: uuid
        offerId:
          type: string
        passengers:
          type: array
          items:
            $ref: '#/components/schemas/PassengerInput'
        payment:
          type: object
          nullable: true
          description: Payment details (PoC — not validated)
    PassengerInput:
      type: object
      required: [firstName, lastName, type]
      properties:
        type:
          type: string
          enum: [ADT, CHD, INF]
        firstName:
          type: string
        lastName:
          type: string
        gender:
          type: string
          enum: [M, F]
        dateOfBirth:
          type: string
          format: date
        nationality:
          type: string
        documentType:
          type: string
        documentNumber:
          type: string
        documentExpiry:
          type: string
          format: date
        documentIssuingCountry:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
    OrderResponse:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
        ndcOrderId:
          type: string
        bookingReference:
          type: string
          example: AB12CD
        airlineCode:
          type: string
        status:
          type: string
          enum: [CONFIRMED, PENDING_PAYMENT, CANCELLED, FAILED]
        totalPrice:
          $ref: '#/components/schemas/Price'
        passengers:
          type: array
          items:
            type: object
            properties:
              passengerId:
                type: string
              type:
                type: string
              firstName:
                type: string
              lastName:
                type: string
        segments:
          type: array
          items:
            $ref: '#/components/schemas/Segment'
        tickets:
          type: array
          nullable: true
          items:
            type: object
    OrderSummary:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
        ndcOrderId:
          type: string
        bookingReference:
          type: string
        airlineCode:
          type: string
        status:
          type: string
        totalPrice:
          $ref: '#/components/schemas/Price'
        segments:
          type: array
          items:
            $ref: '#/components/schemas/Segment'
        createdAt:
          type: string
          format: date-time
...
    AgentProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Verification session ID
        did:
          type: string
          description: Agent DID
        name:
          type: string
        email:
          type: string
        agency:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            iataNumber:
              type: string
        roles:
          type: array
          items:
            type: string